+++ test/NeptunePhysicsTest/npAlignedArray_Test.cpp
@@ -1,8 +1,5 @@
 #include "gtest\gtest.h"
-
 #include "neptunePhysics/core/npAlignedArray.hpp"
-
-#include <iostream>
 
 using namespace NeptunePhysics;
 
@@ -24,13 +21,13 @@ TEST(npAlignedArray, FILL_AND_FETCH_DATA) {
 	npAlignedArray<int> a;
 	a.reserve(20);
 
-	for (size_t i = 0; i < 20; i++)
+	for (int i = 0; i < 20; i++)
 	{
 		a.push_back(i);
 	}
 
 
-	for (size_t i = 0; i < 20; i++)
+	for (int i = 0; i < 20; i++)
 	{
 		EXPECT_EQ(a[i], i);
 		EXPECT_EQ(a.at(i), i);
@@ -50,14 +47,14 @@ TEST(npAlignedArray, FILL_AND_SORT) {
 	npAlignedArray<int> a;
 	a.reserve(20);
 
-	for (size_t i = 0; i < 20; i++)
+	for (int i = 0; i < 20; i++)
 	{
 		int value = i % 3 ? i * (-1) : i * 2;
 		a.push_back(value);
 	}
 
 	a.bubbleSort(compare);
-	for (size_t i = 1; i < 19; i++)
+	for (int i = 1; i < 19; i++)
 	{
 		EXPECT_GT(a[i], a[i - 1]);
 	}
@@ -74,7 +71,7 @@ TEST(npAlignedArray, CHANGE_VALUES_WITH_POINTER) {
 	npAlignedArray<testStruct> a;
 	a.reserve(20);
 
-	for (size_t i = 0; i < 20; i++)
+	for (int i = 0; i < 20; i++)
 	{
 		int value = i;
 		testStruct ts = testStruct(value);
@@ -84,7 +81,7 @@ TEST(npAlignedArray, CHANGE_VALUES_WITH_POINTER) {
 	testStruct* ref = &a.at(19);
 	ref->value = 21;
 
-	for (size_t i = 1; i < 18; i++)
+	for (int i = 1; i < 18; i++)
 	{
 		EXPECT_EQ(a[i].value, i);
 	}
